fix(MoreButton): 필요없는 버튼 삭제
